Updates to secondary nav container
Content template side nav container is still being worked on. Set up to accept the same css rules that the extended nav container from the homepage uses. Issues still exists with text not behaving as it should (i.e. staying on one line and being centered in the span 1 of 4).